과목 등록/조회/삭제 API 구현

+++ src/main/java/com/mathflat/parkdingco/service/RemoveStudentServiceImpl.java
@@ -5,6 +5,8 @@ import com.mathflat.parkdingco.repository.StudentRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
 @Service
 public class RemoveStudentServiceImpl implements RemoveStudentService {
     private final StudentRepository studentRepository;
@@ -18,6 +20,10 @@ public class RemoveStudentServiceImpl implements RemoveStudentService {
     @Override
     @Transactional
     public void remove(Long studentId) {
+        if (Objects.isNull(studentId)) {
+            return;
+        }
+
         scoreRepository.deleteByStudentId(studentId);
 
         if (studentRepository.findById(studentId).isPresent()) {
